[refactor]: 디렉터리 분리 및 리팩토링
